SOLVE THE FIBONACCI PROBLEM
We have implemented not two, but three possible solutions to the PROBLEM
1. The iterative method self.fibs(num) uses just an iterative block.
I'ts time complexity is O(n)

2. The recursive method self.fibs_rec(num) uses a uses an recursive private function to calculate the Fibbonaci number,
but an iterative block to store it in the matrix.
It's code maybe is the most clear, but also the least efficient, because use an recursive method nested
in a iterative loop (I think its an O(n * log n) method)

3. We have also made a pure recursive method self.fibs_pure_rec(num) to solve the problem.
This method uses three arguments: the number, the counter and the response array
When the user calls this method he MUST use just one argument.
¡OH! ¡I THOUGHT THE SOLUTION RIGHT NOW!